requirements.txt
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
statsmodels==0.14.0
pyarrow==12.0.1
openpyxl==3.1.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.0
nltk==3.8.1
flask==2.3.0
fastapi==0.100.0
uvicorn==0.23.0
gunicorn==21.2.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
requests==2.31.0
urllib3==2.0.4
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.4.1
python-dotenv==1.0.0
pyyaml==6.0
jupyter==1.0.0